openarcFlask/requirements.txt
blinker==1.4
braintree==4.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.
eventlet==0.17.4
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.25.0
geopy==2.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
PyJWT==1.6.4
PyMySQL==0.10.0
python-dotenv==0.14.0
python-http-client==3.2.7
requests==2.24.0
sendgrid==6.4.7
six==1.15.0
SQLAlchemy==1.3.18
starkbank-ecdsa==1.0.0
stripe==2.50.0
urllib3==1.25.10
Werkzeug==1.0.1
